create command manager class for util
